Crate gazebo_derive

Derivations for the Gazebo library. Usually you would use these derivations via exports from that library.

Derive Macros

Default_

Derive the Default trait, but without requiring all type arguments to implement Default.

UnpackVariants

VariantName

Derive the VariantName trait.